Backend/requirements.txt
django>=5.2,<6.1
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
django-cors-headers>=4.7,<4.8
Pillow>=11.0,<12.0
